WIP, now shows all pins for all users when a user is not logged in
Currently when a user is logged in, only their own pins will be shown. This will be changed so all pins will be shown with either some way of turning off other peoples pins or displaying them with a different coloured pin. For this I will have to add the user uid to the pin data object